# Conflicts:
#	src/main/java/gregtech/api/capability/impl/AbstractRecipeLogic.java
#	src/main/java/gregtech/api/capability/impl/BoilerRecipeLogic.java
#	src/main/java/gregtech/api/capability/impl/MultiblockFuelRecipeLogic.java
#	src/main/java/gregtech/api/capability/impl/MultiblockRecipeLogic.java
#	src/main/java/gregtech/api/capability/impl/RecipeLogicEnergy.java
#	src/main/java/gregtech/api/metatileentity/SteamMetaTileEntity.java
#	src/main/java/gregtech/api/metatileentity/WorkableTieredMetaTileEntity.java
#	src/main/java/gregtech/api/metatileentity/multiblock/FuelMultiblockController.java
#	src/main/java/gregtech/api/recipes/RecipeBuilder.java
#	src/main/java/gregtech/common/metatileentities/multi/electric/MetaTileEntityDistillationTower.java
#	src/main/java/gregtech/common/metatileentities/multi/electric/MetaTileEntityProcessingArray.java
#	src/main/java/gregtech/loaders/recipe/BatteryRecipes.java
#	src/main/java/gregtech/loaders/recipe/VanillaStandardRecipes.java
#	src/main/java/gregtech/loaders/recipe/chemistry/ChemicalBathRecipes.java
#	src/main/java/gregtech/loaders/recipe/chemistry/MixerRecipes.java
#	src/main/java/gregtech/loaders/recipe/chemistry/PolymerRecipes.java
#	src/main/java/gregtech/loaders/recipe/handlers/MaterialRecipeHandler.java
#	src/main/java/gregtech/loaders/recipe/handlers/PartsRecipeHandler.java
#	src/main/java/gregtech/loaders/recipe/handlers/PipeRecipeHandler.java
#	src/main/java/gregtech/loaders/recipe/handlers/ToolRecipeHandler.java
#	src/main/java/gregtech/loaders/recipe/handlers/WireRecipeHandler.java
#	src/test/java/gregtech/api/capability/impl/AbstractRecipeLogicTest.java
#	src/test/java/gregtech/api/capability/impl/MultiblockRecipeLogicTest.java
#	src/test/java/gregtech/api/recipes/RecipeMapTest.java
#	src/test/java/gregtech/api/recipes/logic/IParallelableRecipeLogicTest.java
#	src/test/java/gregtech/api/recipes/logic/ParallelLogicTest.java

+++ src/main/java/gregtech/api/recipes/builders/AssemblyLineRecipeBuilder.java
@@ -143,7 +143,7 @@ public class AssemblyLineRecipeBuilder extends RecipeBuilder<AssemblyLineRecipeB
         private final ItemStack dataStack;
         private final boolean ignoreNBT;
         private final int duration;
-        private final int EUt;
+        private final long EUt;
         private final int CWUt;
 
         /**
@@ -155,10 +155,10 @@ public class AssemblyLineRecipeBuilder extends RecipeBuilder<AssemblyLineRecipeB
          * @param CWUt          how much computation per tick this recipe needs if in Research Station
          *                      <p>
          *                      By default, will ignore NBT on researchStack input. If NBT matching is desired, see
-         *                      {@link #ResearchRecipeEntry(String, ItemStack, ItemStack, boolean, int, int, int)}
+         *                      {@link #ResearchRecipeEntry(String, ItemStack, ItemStack, boolean, int, long, int)}
          */
         public ResearchRecipeEntry(@NotNull String researchId, @NotNull ItemStack researchStack,
-                                   @NotNull ItemStack dataStack, int duration, int EUt, int CWUt) {
+                                   @NotNull ItemStack dataStack, int duration, long EUt, int CWUt) {
             this.researchId = researchId;
             this.researchStack = researchStack;
             this.dataStack = dataStack;
@@ -177,7 +177,7 @@ public class AssemblyLineRecipeBuilder extends RecipeBuilder<AssemblyLineRecipeB
          * @param CWUt          how much computation per tick this recipe needs if in Research Station
          */
         public ResearchRecipeEntry(@NotNull String researchId, @NotNull ItemStack researchStack,
-                                   @NotNull ItemStack dataStack, boolean ignoreNBT, int duration, int EUt, int CWUt) {
+                                   @NotNull ItemStack dataStack, boolean ignoreNBT, int duration, long EUt, int CWUt) {
             this.researchId = researchId;
             this.researchStack = researchStack;
             this.dataStack = dataStack;
@@ -210,7 +210,7 @@ public class AssemblyLineRecipeBuilder extends RecipeBuilder<AssemblyLineRecipeB
             return duration;
         }
 
-        public int getEUt() {
+        public long getEUt() {
             return EUt;
         }
